Migrate to STM32H743 platform and restructure project for multi-platform support
  Major Changes:
  - Migrate primary target from Raspberry Pi Pico to STM32H743VIT6 (WeAct Mini)
  - Add ST7796S 4" 320×480 display support with SPI interface
  - Reorganize codebase into platform-specific directories (pico/, stm32/, desktop/)
  - Update desktop build system with proper resolution handling

  Platform Structure:
  - platforms/pico/: Original Pico hardware implementation (ST7735 128×160)
  - platforms/stm32/: New STM32H743 target with PlatformIO test projects
  - platforms/desktop/: GLFW/OpenGL development environment for testing

  Desktop Build Improvements:
  - Fix Makefile source paths to work from project root
  - Add platform-specific include paths for cross-platform builds
  - Replace hardcoded display dimensions with DISPLAY_WIDTH/DISPLAY_HEIGHT constants
  - Support dynamic resolution changes through platform_desktop.h
  - Update window and texture creation to use resolution constants

  Hardware Documentation:
  - Add STM32H743 datasheets, schematics, and pinout documentation
  - Add KiCad PCB designs for PocketGate console (4 revisions)
  - Document memory architecture (DTCM, AXI SRAM, SRAM1/2/3/4)
  - Add DMA, QSPI, and memory management guides
  - Include display wiring diagrams and touch controller specs

  Engine Updates:
  - Update framebuffer implementation for larger display (320×480)
  - Enhance fixed-point arithmetic with comparison operators
  - Improve isometric math functions for better performance
  - Update CLAUDE.md with new hardware specs and constraints

  STM32 Drivers:
  - Add ST7796S display driver with DMA support
  - Implement board configuration for WeAct Mini H743VITx
  - Add DFU mode flashing guide
  - Include PlatformIO test projects for driver validation

  Documentation:
  - Move Pico-specific docs to Docs/pico/
  - Add hardware upgrade planning documents
  - Document C++11/14/17/20 standards usage policy
  - Add comprehensive hardware specifications (HARDWARE_SPECS.md)

+++ engine/isometric/iso_math.h
@@ -6,7 +6,7 @@
 #else
     #include "pico/stdlib.h"
 #endif
-#include "fixed_point.h"
+#include "engine/math/fixed_point.h"
 
 struct World_space {
 	// z - up; y - north, x - east
@@ -15,15 +15,14 @@ struct World_space {
 	Fixed_q16 z = 0;
 };
 
-
 struct Screen_space {
-	int x = 0;
-	int y = 0;
+	int16_t x = 0;
+	int16_t y = 0;
 };
 
 struct Camera {
-	int offset_x = 0;
-	int offset_y = 0;
+	int32_t offset_x = 0;
+	int32_t offset_y = 0;
 };
 
 inline Screen_space world_to_screen(const World_space& world, const Camera& cam) {
